// map-app/src/types/game.ts
export interface MarkerInstance {
  id: string;
  category?: string;
  subtype: string;
  region?: string;
  x: number;
  y: number;
  images: string[];
  contributors: string[];
  icon?: string;
  name?: string;
  indexInSubtype: number;
}

export type MarkerWithTranslations = MarkerInstance & {
  localizedName: string;
  localizedDescription?: string;
};